add endpoint list user goals, a user summary, goal detail by user

+++ src/BackEnd/BpChallenge/BpChallenge.Domain/Entities/CurrencyIndicator.cs
@@ -8,9 +8,9 @@ public class CurrencyIndicator : BaseEntity
     public Currency SourceCurrency { get; private set; }
     public int DestinationCurrencyId { get; private set; }
     public Currency DestinationCurrency { get; private set; }
-    public decimal Value { get; private set; }
+    public double Value { get; private set; }
     public DateTime Date { get; private set; }
     public DateTime RequestDate { get; private set; }
-    public decimal Ask { get; private set; }
-    public decimal Bid { get; private set; }
+    public double Ask { get; private set; }
+    public double Bid { get; private set; }
 }
